requirements.txt
flask
flask_login
flask_wtf
mysql-connector-python
